requirements.txt
playwright==1.44.0
anthropic==0.28.0
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.32.3
websocket-client==1.8.0
python-telegram-bot==21.3
